refactor(state-machine): split markup serialization from definitions
