+++ code/packages/rust/Cargo.toml
@@ -323,4 +323,5 @@ members = [
     "window-c",
     "tcp-runtime",
     "udp-client",
+    "state-machine-markup-serializer",
 ]
