+++ code/packages/rust/state-machine-markup-serializer/BUILD
@@ -1,0 +1,1 @@
+cargo test -p state-machine-markup-serializer -- --nocapture
