+++ code/packages/rust/state-machine-markup-serializer/CHANGELOG.md
@@ -1,0 +1,13 @@
+# Changelog
+
+All notable changes to this package will be documented in this file.
+
+## [0.1.0] - 2026-04-20
+
+### Added
+
+- Initial package scaffolding generated by scaffold-generator.
+- Added deterministic `.states.toml` serialization for typed
+  `StateMachineDefinition` values.
+- Added snapshot coverage for DFA, NFA, PDA, epsilon, multi-target, stack-effect,
+  and TOML string escaping behavior.
